+++ lib/rules/color-named/__tests__/colordUtils.test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { colord } = require('../colord-utils');
+const { colord } = require('../colordUtils');
 
 describe('colord gray', () => {
 	const TESTS = [
