+++ lib/rules/color-named/colordUtils.js
