+++ lib/rules/color-named/index.js
@@ -10,7 +10,7 @@ const ruleMessages = require('../../utils/ruleMessages');
 const validateOptions = require('../../utils/validateOptions');
 const valueParser = require('postcss-value-parser');
 const { isRegExp, isString } = require('../../utils/validateTypes');
-const { colord } = require('./colord-utils');
+const { colord } = require('./colordUtils');
 
 const ruleName = 'color-named';
